Move SDP precoders into sims
