+++ sims/SDP/run.py
@@ -1,5 +1,6 @@
 import sys
 import os
+import precoder
 
 # from multiprocessing import Process
 scriptdir = os.path.dirname(os.path.realpath(__file__))
@@ -7,15 +8,15 @@ sys.path.append(os.path.join(scriptdir, '..', '..'))
 sys.path.append(os.path.join(scriptdir, '..', '..', 'beamy'))
 
 from beamy.simulator import Simulator  # noqa: E402
-import precoder  # noqa: E402
+from beamy.precoder import PrecoderWMMSE  # noqa: E402
 
-####
+
 realizations = 25
 biterations = 50
 
 
 def simulate(_rx, _tx, _K, _B, _SNR):
-    sim = Simulator(precoder.PrecoderWMMSE(),
+    sim = Simulator(PrecoderWMMSE(),
                     bs=_B, users=_K, nr=_rx, nt=_tx,
                     realizations=realizations, biterations=biterations,
                     SNR=_SNR)
